Replace old Collection Data Source with the new one
